Create ObjectType1.java
A sales person is paid commission based on the sales he makes as shown by the following
table:

SALES COMMISSION
Under Rs. 500 2% of SALES
Rs. 500 and under Rs. 5000 5% of SALES
Rs. 5000 and over 8% of SALES

Write a class, Commission, which has:
 An instance variable, sales; an appropriate constructor; and a method, getCommission()
that returns the commission.
Now write a Demo class in Java to test the Commission class by reading a sale from the
user, using it to create a Commission object after validating that the value is not negative.
Finally, call the getcommission() method to get and print the commission. If the sales are
negative, your Demo class should print the message “Invalid Input”.